+++ NEWS.md
@@ -1,3 +1,12 @@
+# spNetwork 0.1.2.9000
+
+* Added two function for datadriven bandwidth selectionn (see vignette NKDE)
+* Added a function to simplify network before processing (*simplify_network*), reducing network complexity by removing useless edges or merge continuous edge can improve performance.
+
+# spNetwork 0.1.1
+
+* Accepted CRAN version (minor changes)
+
 # spNetwork 0.1.0
   
 * Initial release version